Add pagination (#22)
* Add support for pagination (#21)

Creates endpoint for pagination and uses offset and limit to get the
items one page at a time.

* Enable pagination support

Update javascript to use the paginated url to retrieve memes. Prevent
new requests when the previous one has not been completed yet, and stop
when an emtpy set is returned (last item has been fetched).

* Make sure scrolling is always possible

Scrolling shoulld always be possible, even when images do not fill the
screen completely. Scrolling is required to trigger loading of more
images.